API productos y carrito

+++ controllers/productsRouter.js
@@ -12,6 +12,12 @@ routerProducts.use(express.json());
 // instanciamos la base de productos
 const productos = new Productos(`./database/productos.json`);
 
+// admin options
+
+let administrador = true;
+
+// requests
+
 routerProducts.get(`/:id?`, (req, res) => {
   if (req.params.id) {
     const product = productos.getById(req.params.id);
@@ -23,18 +29,45 @@ routerProducts.get(`/:id?`, (req, res) => {
 });
 
 routerProducts.post(`/`, (req, res) => {
-  const newProduct = productos.saveProduct(req.body);
-  res.status(200).json(newProduct);
+  if (administrador) {
+    const newProduct = productos.saveProduct(req.body);
+    res.status(200).json(newProduct);
+  } else {
+    res.status(401).json({
+      error: new Error(`Acceso no autorizado`),
+      descripcion: `carga de datos sin permiso`,
+      ruta: req.path,
+      método: req.method,
+    });
+  }
 });
 
 routerProducts.put(`/:id`, (req, res) => {
-  const modifiedProduct = productos.modify(req.params.id, req.body);
-  res.status(200).json(modifiedProduct);
+  if (administrador) {
+    const modifiedProduct = productos.modify(req.params.id, req.body);
+    res.status(200).json(modifiedProduct);
+  } else {
+    res.status(401).json({
+      error: new Error(`Acceso no autorizado`),
+      descripcion: `actualización de datos sin permiso`,
+      ruta: `api/productos/${req.params.id}`,
+      método: `PUT`,
+    });
+  }
 });
 
 routerProducts.delete(`/:id`, (req, res) => {
-  const deletedProduct = productos.deleteById(req.params.id);
-  res.status(200).json(deletedProduct);
+  if (administrador) {
+    const deletedProduct = productos.deleteById(req.params.id);
+    res.status(200).json(deletedProduct);
+  } else {
+    res.status(401).json({
+      error: new Error(`Acceso no autorizado`),
+      descripcion: `eliminación de datos sin permiso`,
+      ruta: `api/productos/${req.params.id}`,
+      método: `DELETE`,
+    });
+  }
 });
 
-module.exports = routerProducts;
+module.exports = { routerProducts: routerProducts, productos: productos };
